Version 2.7 (4. November 2018) - Changed license to LGPLv3 - CMake toolchain support. (buildit.m is deprecated now!) - Separated C-Blosc as optional submodule. mksqlite supports all compression   types now.   C-Blosc is integrated as "submodule" in the mksqlite git repository and   is not part of the mksqlite sources anymore.   You may compile mksqlite without C-Blosc as required, adjustable per   MKSQLITE_CONFIG_USE_BLOSC define. - Updated MD5 source   [http://openwall.info/wiki/people/solar/software/public-domain-source-code/md5] - SQLite update to V3.25.2

+++ value.hpp
@@ -10,9 +10,9 @@
  *             - ValueSQLCol holding a complete table column
  *  @authors   Martin Kortmann <mail@kortmann.de>, 
  *             Andreas Martin  <andimartin@users.sourceforge.net>
- *  @version   2.5
- *  @date      2008-2017
- *  @copyright Distributed under LGPL
+ *  @version   2.7
+ *  @date      2008-2018
+ *  @copyright Distributed under LGPLv3
  *  @pre       
  *  @warning   
  *  @bug       
@@ -501,7 +501,7 @@ public:
     inline
     int NumDims() const
     {
-        return m_pcItem ? mxGetNumberOfDimensions( m_pcItem ) : 0;
+        return m_pcItem ? (int)mxGetNumberOfDimensions( m_pcItem ) : 0;
     }
       
     /**
